add upload overlay

+++ app/client-side/src/services/upload.service.ts
@@ -6,14 +6,21 @@ import { baseUrl } from '.';
 import { FurnitureCardControlService } from './furniture-card-control.service';
 import { ServerImageControlService } from './server-image-control.service';
 import { AccountCookieService } from './account-cookie.service';
+import { NotificationService } from './notification.service';
 
-type UploadStatus = 'uploading' | 'success' | 'error' | 'cancelled';
-
+type uploadStatus = 'uploading' | 'success' | 'error';
+export type uploadType = 'create' | 'update';
 export interface UploadEntry {
   id: string;
   name: string;
   progress: number;
-  status: UploadStatus;
+  status: uploadStatus;
+}
+interface MyXhrUploadOptions {
+  endpoint: string;
+  fieldName?: string;
+  formData?: boolean;
+  getResponseError?: (responseText: string, response: Response) => Error | null;
 }
 
 @Injectable({ providedIn: 'root' })
@@ -25,45 +32,67 @@ export class UploadService {
   constructor(
     private furnitureCardService: FurnitureCardControlService,
     private serverImageControl: ServerImageControlService,
-    private accountCookieService: AccountCookieService
+    private accountCookieService: AccountCookieService,
+    private notification: NotificationService
   ) {
     this.uppy = new Uppy({ autoProceed: true });
 
     this.uppy.use(XHRUpload, {
       endpoint: this.baseServiceUrl,
       fieldName: 'file',
-      formData: true
-    });
+      formData: true,
+      async onBeforeRequest(xhr: any) {
+        console.log('onBeforeRequest', xhr);
+      },
+      shouldRetry: (error: any, response: any) => {
+        console.log(error, response);
+      },
+      getResponseError: (responseText, response) => {
+        console.error('Ответ от сервера:', responseText, response);
+        return new Error(`Ошибка: ${response.statusText}`);
+      }
+    } as MyXhrUploadOptions);
 
     this.uppy.on('upload-progress', (file, progress) => {
-      if (file === undefined || file.name === undefined) return;
-      this.updateUpload(file.id, file.name, progress.percentage ?? 50, 'uploading');
+      console.log('Progress:', progress, file);
+
+      if (file === undefined || file.meta.name === undefined) return;
+
+      const NAME = file.data instanceof File ? file.data.name : 'file'
+      this.updateUpload(file.id, NAME, file.progress.percentage ?? 50, 'uploading');
     });
 
     this.uppy.on('upload-success', (file) => {
-      if (file === undefined || file.name === undefined) return;
-      this.updateUpload(file.id, file.name, 100, 'success');
+      console.log('Success:', file);
+
+      if (file === undefined || file.meta.name === undefined) return;
+
+      const NAME = file.data instanceof File ? file.data.name : 'file'
+      this.updateUpload(file.id, NAME, 100, 'success');
+      this.notification.setSuccess('Модель обновлена', 2500);
     });
 
-    this.uppy.on('upload-error', (file) => {
-      if (file === undefined || file.name === undefined) return;
-      this.updateUpload(file.id, file.name, 0, 'error');
-    });
+    this.uppy.on('upload-error', (file, error) => {
+      console.log('Error:', error.message);
 
-    this.uppy.on('file-removed', (file) => {
-      if (file === undefined || file.name === undefined) return;
-      this.updateUpload(file.id, file.name, 0, 'cancelled');
+      if (file === undefined || file.meta.name === undefined) return;
+
+      const NAME = file.data instanceof File ? file.data.name : 'file'
+      this.updateUpload(file.id, NAME, 100, 'error');
+      this.cancelUpload(file.id);
     });
   }
 
-  addFile(file: File, jwt: string, furnitureCardId: string) {
+  addFile(file: File, jwt: string, furnitureCardId: string, uploadType: uploadType) {
+    const UNIQUE_ID = Date.now().toString();
     this.uppy.addFile({
-      name: file.name,
+      name: UNIQUE_ID + this.getFileExtension(file.name),
       type: file.type,
       data: file,
       meta: {
         furnitureCardId: furnitureCardId,
-        jwt: jwt
+        jwt: jwt,
+        uploadType: uploadType
       }
     });
   }
@@ -74,23 +103,27 @@ export class UploadService {
       this.uppy.removeFile(fileID);
       const JWT = this.accountCookieService.getJwt()
       const FURNITURE_CARD_ID = file.meta['furnitureCardId'] as string;
+      const UPLOAD_TYPE = file.meta['uploadType'] as string;
 
-      this.removeFileFromQueue(fileID)
 
-      if (JWT&&FURNITURE_CARD_ID) {
-        this.serverImageControl.DELETEproject(JWT, FURNITURE_CARD_ID)
+      if (JWT && FURNITURE_CARD_ID && UPLOAD_TYPE === 'create') {
+        await this.serverImageControl.DELETEproject(JWT, FURNITURE_CARD_ID)
         this.furnitureCardService.DELETEfurnitureCard(JWT, FURNITURE_CARD_ID)
       }
     }
   }
-
+  private getFileExtension(fileName: string): string | null {
+    const parts = fileName.split('.');
+    const EXTENSION = parts.length > 1?parts.pop()?.toLowerCase():"obj"
+    return `.${EXTENSION}`
+  }
   removeFileFromQueue(fileID: string) {
     const updatedQueue = this.uploads$.value.filter(file => file.id !== fileID);
     this.uploads$.next(updatedQueue);
   }
 
-  private updateUpload(id: string, name: string, progress: number, status: UploadStatus) {
-    const current = this.uploads$.value.filter(u => u.id !== id);
+  private updateUpload(id: string, name: string, progress: number, status: uploadStatus) {
+    const current = this.uploads$.value.filter(file => file.id !== id);
     this.uploads$.next([...current, { id, name, progress, status }]);
   }
 }
